+++ toglacier.go
@@ -4,43 +4,19 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"strconv"
 	"strings"
 
 	"github.com/jasonlvhit/gocron"
 	"github.com/rafaeljusto/toglacier/internal/archive"
 	"github.com/rafaeljusto/toglacier/internal/cloud"
+	"github.com/rafaeljusto/toglacier/internal/config"
 	"github.com/rafaeljusto/toglacier/internal/storage"
 	"github.com/urfave/cli"
 )
 
-var awsAccountID, awsVaultName, auditFile string
-var backupPaths []string
-var keepBackups = 10
-
-func init() {
-	var err error
-
-	awsAccountID = os.Getenv("AWS_ACCOUNT_ID")
-	awsVaultName = os.Getenv("AWS_VAULT_NAME")
-	backupPaths = strings.Split(os.Getenv("TOGLACIER_PATH"), ",")
-	auditFile = os.Getenv("TOGLACIER_AUDIT")
-
-	if os.Getenv("TOGLACIER_KEEP_BACKUPS") != "" {
-		if keepBackups, err = strconv.Atoi(os.Getenv("TOGLACIER_KEEP_BACKUPS")); err != nil {
-			fmt.Printf("invalid number of backups to keep. details: %s", err)
-			os.Exit(1)
-		}
-	}
-}
-
 func main() {
-	awsCloud, err := cloud.NewAWSCloud(awsAccountID, awsVaultName, false)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-	auditFileStorage := storage.NewAuditFile(auditFile)
+	var awsCloud cloud.Cloud
+	var auditFileStorage storage.Storage
 
 	app := cli.NewApp()
 	app.Name = "toglacier"
@@ -52,12 +28,43 @@ func main() {
 			Email: "adm@rafael.net.br",
 		},
 	}
+	app.Flags = []cli.Flag{
+		cli.StringFlag{
+			Name:  "config, c",
+			Usage: "Tool configuration file (YAML)",
+		},
+	}
+	app.Before = func(c *cli.Context) error {
+		config.Default()
+
+		var err error
+
+		if c.String("config") != "" {
+			if err = config.LoadFromFile(c.String("config")); err != nil {
+				fmt.Printf("error loading configuration file. details: %s\n", err)
+				return err
+			}
+		}
+
+		if err = config.LoadFromEnvironment(); err != nil {
+			fmt.Printf("error loading configuration from environment variables. details: %s\n", err)
+			return err
+		}
+
+		if awsCloud, err = cloud.NewAWSCloud(config.Current(), false); err != nil {
+			fmt.Printf("error initializing AWS cloud. details: %s\n", err)
+			return err
+		}
+
+		auditFileStorage = storage.NewAuditFile(config.Current().AuditFile)
+		return nil
+	}
 	app.Commands = []cli.Command{
 		{
 			Name:  "sync",
 			Usage: "backup now the desired paths to AWS Glacier",
 			Action: func(c *cli.Context) error {
-				backup(awsCloud, auditFileStorage)
+				backup(config.Current().Paths, awsCloud, auditFileStorage)
 				return nil
 			},
 		},
@@ -109,8 +116,8 @@ func main() {
 			Usage: "run the scheduler (will block forever)",
 			Action: func(c *cli.Context) error {
 				scheduler := gocron.NewScheduler()
-				scheduler.Every(1).Day().At("00:00").Do(backup, awsCloud, auditFileStorage)
-				scheduler.Every(1).Weeks().At("01:00").Do(removeOldBackups, awsCloud, auditFileStorage)
+				scheduler.Every(1).Day().At("00:00").Do(backup, config.Current().Paths, awsCloud, auditFileStorage)
+				scheduler.Every(1).Weeks().At("01:00").Do(removeOldBackups, config.Current().KeepBackups, awsCloud, auditFileStorage)
 				scheduler.Every(4).Weeks().At("12:00").Do(listBackups, true, awsCloud, auditFileStorage)
 				<-scheduler.Start()
 				return nil
@@ -122,7 +129,7 @@ func main() {
 			Usage:     "encrypt a password or secret",
 			ArgsUsage: "<password>",
 			Action: func(c *cli.Context) error {
-				if pwd, err := cloud.PasswordEncrypt(c.Args().First()); err == nil {
+				if pwd, err := config.PasswordEncrypt(c.Args().First()); err == nil {
 					fmt.Printf("encrypted:%s\n", pwd)
 				} else {
 					fmt.Printf("error encrypting password. details: %s\n", err)
@@ -135,7 +142,7 @@ func main() {
 	app.Run(os.Args)
 }
 
-func backup(c cloud.Cloud, s storage.Storage) {
+func backup(backupPaths []string, c cloud.Cloud, s storage.Storage) {
 	archive, err := archive.Build(backupPaths...)
 	if err != nil {
 		log.Println(err)
@@ -232,7 +239,7 @@ func removeBackup(id string, c cloud.Cloud, s storage.Storage) {
 	}
 }
 
-func removeOldBackups(c cloud.Cloud, s storage.Storage) {
+func removeOldBackups(keepBackups int, c cloud.Cloud, s storage.Storage) {
 	backups := listBackups(false, c, s)
 	for i := keepBackups; i < len(backups); i++ {
 		removeBackup(backups[i].ID, c, s)